:honeybee: (svg-tester) rename file
