+++ devTools/svgTester/chart-configurations.ts
